WS-008: Move create user function to api client file
